add path config "luci"
infra_paths module defines custom paths (e.g. git_cache or slave_build)
that are used by various recipes and recipe modules (e.g. gclient). The
problem is that either infra_paths configurations have too much
knowledge about build environment or chromium project.

In LUCI all we need is cwd and cache dirs. Both are defined in path
module, so we don't need infra_paths to do anything. We cannot easily
remove infra_paths module because other recipe modules, e.g. gclient,
statically depend on it. Also we cannot make it do nothing if
path_config property is not passed because it would cause all recipe
expectation files in the recipe universe to change.

Instead we introduce a new path config called "luci" which tells
infra_paths module to not register any new paths. It is equivalent to
not using infra_paths module.

Currently gclient and bot_update modules fail if slave_build or git_cache paths
are not defined. Change that so they fail back to luci-standard paths:
- [GIT_CACHE] -> [CACHE]/git
- [SLAVE_BUILD] -> [CWD]

Note that [CACHE] ad [CWD] are defined in path recipe module and that
the value of [CACHE] can be set via a property and defaults to
[CWD]/cache.

R=iannucci@chromium.org, martiniss@chromium.org
BUG=662586