remove slave_build path
slave_build path, defined by infra_paths module, is unnecessarily tied
to our buildbot setup which we are trying to deprecate in favor of LUCI.
In practice it is always == CWD and that's what we want in LUCI too.

Update all code that uses slave_build path to use cwd instead.

R=iannucci@chromium.org, martiniss@chromium.org
BUG=662586

Change-Id: Id52aa535d6a4f52670d581bf56aab65f670b1861