remove slave_build path
slave_build path, defined by infra_paths module, is unnecessarily tied
to our buildbot setup which we are trying to deprecate in favor of LUCI.
In practice it is always == start_dir and that's what we want in LUCI too.

Update all code that uses slave_build path to use start_dir instead.

R=iannucci@chromium.org, martiniss@chromium.org
BUG=662586