Update patch set 3
Patch Set 3:

I think it doesn't fully work for us. What works it that the v8 revision on the tester is the on from the builder, done by this feature:
https://cs.chromium.org/chromium/build/scripts/slave/recipe_modules/chromium_tests/api.py?l=164

For this we set the set_component_rev property on our bots, e.g.:
https://cs.chromium.org/search/?q=set_component_rev&sq=package:chromium&type=cs

The TODO in the code above is what's missing to also let chromium sync to the builder's revision.

Patch-set: 3