Update patch set 7
Patch Set 7:

Yes, it is pinned separately. But it is available in PATH only after running 'env.py'. Now that cipd is in PATH via depot_tools, I'm worried we'll be using inconsistent versions of the client (since what version is used is now depends on whether env.py was used).

Patch-set: 7