Update patch set 1
Patch Set 1: Code-Review+1

LGTM if it works (please verify it works).

It reminded me of an issue we had in Go toolset bootstrap: if multiple concurrent scripts happen to call unbootstrapped depot_tools/cipd at the same time, there'll be a race in bootstrapping it. It can fail on Windows.

Patch-set: 1
Label: Code-Review=+1