Update patch set 5
Patch Set 5:

I'm still confused by this change.

It looks like in the old version of the code, if there were errors this did nothing. In the new code, we call output.fail(). Is that call necessary? If so, would this mean things were actually broken before?

Second, I'm not seeing how the "there were warnings but no errors" logic changes at all?

Third, I was suggesting that if someone passes `-f`, we should ignore *both* errors and warnings. I kinda though agable agreed with that, but that's not what this change does, right?

Patch-set: 5