Update patch set 5
Patch Set 5:

(2 comments)

> Patch Set 5:
> 
> I'm still confused by this change.
> 
> It looks like in the old version of the code, if there were errors this did nothing. In the new code, we call output.fail(). Is that call necessary? If so, would this mean things were actually broken before?

Good point, it does look like if there are errors then output.fail() wouldn't have been called there in the original, but would be called there now. This change was not intended :-/

> Second, I'm not seeing how the "there were warnings but no errors" logic changes at all?

It's harder to see since I moved some other stuff around, but in the "there were warnings but no errors and may_prompt is False" case, then I think in the original output.fail() should be called, and in the changed version, output.fail() will not be called.

> Third, I was suggesting that if someone passes `-f`, we should ignore *both* errors and warnings. I kinda though agable agreed with that, but that's not what this change does, right?

Right, that's not what this change does. This change still keeps it so that if there are presubmit errors, we'd still fail even if -f is passed.

Patch-set: 5