Rip git-svn code out of roll-dep-svn
This CL feels rather ridiculous. The whole point of the "-svn" in the
name of this script is that it can handle SVN revisions, but will
still put a git hash in the DEPS file. However, we no longer have any
SVN deps to roll, so that functionality is useless. However however,
the functionality of roll-dep-svn is sufficiently different from the
plain roll-dep script that we can't simply delete it and migrate all
its users over. Hence, this CL: remove all the SVN code, let the
current users keep using it unchanged, and figure out something else
saner later.

BUG=642523

Change-Id: I2ce005c7bb5427558463fb70e57f3b27c0a2dbb2