Update patch set 1
Patch Set 1: Code-Review+1

(3 comments)

Thanks for taking care of this. I have a comment and two suggestions below. I'd have no problem moving the remaining code to v8 and let it get rusty there. But in depot_tools is a good place ATM to share this kind of code.

Patch-set: 1
Label: Code-Review=+1