Update patch set 2
Patch Set 2:

> > I believe this forces me to format .js files under
 > > third_party/WebKit/LayoutTests when I change any single line.
 > 
 > fwiw: I didn't mean to commit this, I just never use gerrit.
 > 
 > Dmitry: are you experiencing bad formatting in your layout tests? 
 > basically I'm wondering if you're looking to turn off formatting
 > those files because you don't like the formatting or because you
 > believe those files shouldn't be formatted?  also, aren't most
 > layout tests in .html files?

I believe blink has all kinds of styling and formatting in layout tests, and IIRC there was a thread on blink-dev where many agreed we should not force any particular style for layout tests.

Most tests are html, but there are a lot of script resources which are .js files.

Patch-set: 2
Label: Code-Review=0