Update patch set 2
Patch Set 2:

(1 comment)

The fact that this CL also contains refactoring makes it really hard to confirm that this is also reducing the number of RPCs. I'd expect to have the test diff in this CL be simply removing a bunch of lines, but instead there are a bunch of insertions and it is hard to know that this is correct.

Even though the CL is small, would you mind splitting the refactoring into a predecessor CL?

Patch-set: 2
Label: Code-Review=0