Change canned presubmit check to respect 100 character limit for ObjC.
A recent change in clang-format made it respect the 100 characters
limit for source and header files for Objective-C which causes the
CL to fail at presubmit.

Change canned presubmit check to respect the 100 characters limit
for file with 'm' or 'mm' and to also respect it for 'h' file if
there is a corresponding Objective-C implementation file.

BUG=686650

Change-Id: I9ce9b1224cc4b175985bb921a1e74e79139e40cd