Update patch set 2
Patch Set 2: Code-Review+1

Really, this is all that's necessary? That's great.

Please please *please* fix recipe_modules.rietveld.calculate_issue_root. It has its own map, which is not as complete as this one. It appears to only be called by recipe_modules.tryserver, and then only in specific circumstances.

I spent a good few hours yesterday trying to figure out the best way to generalize it and make sure it would work for gerrit and make sure that it was all that was necessary. Turns out it wasn't necessary at all.

At the very least, document it. Preferably, remove it. I doubt anyone is actively using it.

Patch-set: 2
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>