Update patch set 1
Patch Set 1:

If this looks good to hinoka@, I'm fine with it as well, but I'd like to know how we got into the state that files were left lying around. Unfortunately, when I tried to look at the logs on the builder in question, things were messed up and I couldn't actually see what had happened in prior builds.

Patch-set: 1