Do not run unconditionally run pre_deps_hooks for dependencies.
Currently if you set `pre_deps_hooks` in your gclient DEPS file,
they will be executed regardless of whether your repo was supposed
to be recursed into (via recursionlimit or recursedeps) or not.

This change fixes that so that we only run the hooks that were
explicitly configured to be run.

R=kjellander@chromium.org, agable@chromium.org
BUG=696495

Change-Id: Ic0ef641903cf3f56d7c2e119d8b0db862258d995