Workaround Python Issue 26083 in upload.py
This issue is fixed in Python 2.7.12, but we don't guarantee
that we use that version everywhere. The root issue is that,
on certain platforms (e.g. Mac), the exception indicating that
the underlying command (e.g. 'hg') was not found is too large,
and loading the pickle representing the exception fails.

This CL catches this specific exception and treats it identically
to the OS properly raising a 'command not found' code.

BUG=699173

Change-Id: I0483ab0d1d0b6a3dfcfd26f2d9644817b5b5284f