define builder_cache base path
chromium_tests recipe module currently checks if builder_cache base path
is defined, and if it is not it let's the caller to decide where to put
the cache. All the callers choose start_dir.

In LUCI we want to use [CACHE]/builder as builder cache. To do that we
want to update chromium_tests recipe module to use [CACHE]/builder if
builder_cache is not defined.

Define builder_cache in buildbot to point to start dir, so that in LUCI
it is not defined and thus [CACHE]/builder is used.

R=iannucci@chromium.org
BUG=660481

Change-Id: I9018da5c43a5396add6a26ad2c12b1755fbe6307