Update packaging script to support VS 2017
VS 2017 has been released and needs to be packaged so that we can
experiment with building with it. This is updates the packaging
script to add support for VS 2017 while removing support for VS 2013.
The file layout has changed significantly in VS 2017. Compatibility with
VS 2015 has, I believe, been maintained but it is not important enough
to merit significant testing.

Packages built with this script have been used to build Chrome.

BUG=683729

Change-Id: I68e5a8d9fd389132b641743dbc070108497f54cb