Reland: add "generic" infra path config
It proved to be dangerous to introduce conditional logic into recipes
and recipe modules to alternate between buildbot and luci modes, namely
it touches buildbot code paths, which can be very dangerous.

Instead, introduce "generic" path config that configures
existing base paths from the built-in ones.

Also remove swarmbucket path config because it is not used.

--

This CL is a simpler version of 
https://chromium-review.googlesource.com/c/459234/
in that it does not make generic path config a default.
This avoids behavioral change in downstream recipe packages and 
makes rolling into build.git trivial.

R=iannucci@chromium.org
BUG=660481

Change-Id: Ifc147e29e68a6e0c800da73a8e356238dfc2938a