Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 2: Code-Review+1 Commit-Queue+2
> > 
> > No need to flag again; you can see at the top that the CR+1 from maruel@ is "sticky" and still applies to your rebase patchset. Submitting to CQ.
> 
> Even with LGTM status I still can't make bot commit it. This message editing form says:
> 
>   Code-Review  You don't have permission to edit this label.
>   Commit-Queue You don't have permission to edit this label.

Yep, that's because the ability to set the Commit-Queue label is the same as the ability to run (potentially untrusted) code on our infrastructure. So we don't grant it to just anyone. But I've nominated you for tryjob access, so the next time you upload an infra CL you should be able to set the CQ label yourself!

Patch-set: 3