Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > I would prefer that "git map -h" just be an alias for "git help map", which is a full man page and the canonical documentation. Having two different sources of documentation is asking for trouble.
> 
> This is not how all other commands work. Also, it "help map" requires visual access to browser windows, which is not available in Linux containers I usually work on.

It does not require browser access. If you're running on "git help map" on linux, it will pull up an actual man page in your current terminal emulator, not an html page in your browser.

Thanks for filing that help-broken-on-windows bug and the fix for it.

Patch-set: 3
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>