Avoid updating same dep multiple times
DO_NOT_SUBMIT: I realize this is not safe to land, since it will
break the case when the same dep path is used for multiple platforms
but not using the same repo path and/or revision.

This avoids errors when lss_revision is about to be updated.
Because this dep is present in multiple deps_os entries, it
becomes updated multiple times, which the code is not handling well,
resulting in an invalid DEPS file.

Bug: 708896
Change-Id: Iab9124ee53d908750182b9880ce8e6b3870025dd