Update patch set 2
Patch Set 2:

Unfortunately we still aren't getting crash dumps. A pdfium builder hit a linker crash today so I VNC connected to the builder. I confirmed that the registry key is created correctly (and I also saw the old 32-bit registry key) but no crash dump was created. On Windows 10 just creating the registry key is sufficient. Maybe I need to create the directory? Create the sub-keys? Give up and become a barista? I even did local testing on a build machine with an intentionally crashing executable and with the subdirectory and sub-keys created, but no dice.

I also did a local build of the target that the linker crashed when building to verify that we are using /fastfail. We are.

Maybe this is flaky on Windows 7? I just don't know, but I'm giving up for now.

Patch-set: 2