Update patch set 4
Patch Set 4:

Hm.. Why are you trying to avoid powershell? Part of removing the python bootstrap scripts to replace them with CIPD was going to be to remove the js downloader script as well. Powershell is much better supported than wscript afaik, which is why I did it that way.

The goal was to make the bootstrap scripts for cipd as short and as similar as possible, which is why I chose bash and powershell.

Patch-set: 4
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>