[presub] lstrip line limit exceptions; ignore <g and <path
The <link lines in HTML are allowed to have longer than 80 characters.
This CL adds <g and <path to the list of unlimited characters on a line.
(These g and path tags are used in svg image data).

These new tags may be preceded by blank space so .lstrip() was added to
the test. I believe this .lstrip() should not be an issue for the other
languages and exemptions present.

Bug: None
Change-Id: Ie9c2bca0e7147033e360ddcfc6eee3b73e167228