Update patch set 4
Patch Set 4:

re: repo, that's sort-of essentially what we're planning to turn depot_tools into: download a tool (maybe `depot_tool.exe` :)), which knows how to keep itself up to date and how to pull/verify all packages (e.g. like the git commands).

At that point the git commands would indeed be an independent package (I want this for myself :)). There would also be a tool-less way to obtain the packages (i.e. with just a browser), and we could (finally) separate depot_tools into a couple well defined pieces (maybe putting the separate pieces of source in different folders! Imagine that! :)) instead of the giant mashup it is today.

That said, my current not-a-guarantee-at-all guesstimate for when this might happen would be sometime in Q3 at the earliest.

Patch-set: 4