Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I believe all the versions of Linux we care about already have a version of coreutils that's aware of --reflink, so you could just pass --reflink=auto on linux and not have any of these options I think?

It's true that we can use --reflink=auto, but I think it makes more sense to separate the two use cases.

In this CL, with CoW support, all the files are copied, including files downloaded by gclient hook, artifacts, and all the other files not in .gitignore. Furthermore, the way it's implemented is to copy everything first, and then delete the .git folders and use "git clean".

On non-CoW FS, the cost of copying everything first is much higher, and it's very easy to waste disk space on things we don't really need or care. Symlinking the .git and do "git checkout" would result in a clean folder, and I think that should still be the way to go on non-CoW FS.

Patch-set: 5
Reviewer: Gerrit User 1144344 <1144344@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>