Update patch set 1
Patch Set 1: Code-Review+1

This feels like the wrong solution -- even though the change is uploaded in --no-squash mode, that doesn't mean that we shouldn't be adding CCs from WATCHLISTS, or that we should drop the -r and --cc flags on the floor. A better solution would be to (in git_cl.py) get the change IDs of all the CLs which were just created, and call AddReviewers once for each of them.

That said, --no-squash is officially unsupported. So I guess it can be as broken as you like. I'll approve this, but if you have the time to look into what I proposed above I'd take that as a great favor.

Patch-set: 1
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1