Remove special Cygwin PATH manipulation from .bat.
Several boilerplate batch files include a provision to prepend
"depot_tools" to PATH prior to running those tools. This undermines
the utility of PATH overrides, since these tools specifically force their
"depot_tools" sub-paths to be used regardless of environment.

The origin of this behavior is likely limited to a specific fix for a
specific problem, but was then perpetuated by the copy/paste of
boilerplate bootstrap code as more bootstraps were added.

This is important in upcoming configurations, where core tools such as
Python and Git will be overridden via PATH on the bots.

Cygwin users who depended on this behavior should just add "depot_tools"
to their PATH in the appropriate location (i.e. in their .bashrc).

BUG=chromium:714293
TEST=None

Change-Id: Ie948a430847d20326d2411e9296cacd02f83a537