[.bat] Remove "depot_tools" override.
Currently, all ".bat" entry points use "~dp0python" to ensure that the
Python that is used to execute the tool is the one in depot_tools. This
prevents any sort of system override.

Remove this override so that PATH solely determines which Python is
used. To accommodate users who invoked these tools without Python on the
PATH, we still still add "depot_tools" as a catch-all PATH suffix.

BUG=chromium:714293, chromium:724902
TEST=None

Change-Id: I06e9583a668c767196a2a335547aded868f2a2b5