[git] Simplify checkout logic.
This is a cleanup CL aimed to break apart the fairly-heavy "git"
module's "checkout" method into functional components and a backing
script. This isn't a complete cleanup, but it brings the module more
in-line with somewhere we can be comfortable using it for critical
things.

Notably, this adds "git_util.py", which performs Git operations. This
has a nice side-effect of being a single output log, which will make
output parsing a bit easier and allow for some more verbose and/or
introspective output options (e.g., dump Git version). "git_util"
currently supports:
- "setup", which is given a directory and sets it up as a Git repository
  with a remote, ready for future operations.
- "ensure_checkout", which performs a "setup" and ensures that the
  ensuing repository is checked out to the appropriate ref.

The script has room to grow for other operations, too.

We also remove some "checkout" keyword arguments, since codesearch
suggests that they aren't actually used anywhere. The submodule
functionality has been rolled into a simple "--recurse" argument, and
specific file checkout, if encountered, will be combination "setup" /
"checkout -- <file>" call instead of a special case of a more generic
setup function.

The "can_fail_build" argument has also been removed, as it doesn't
appear to be in use. Additionally, though, it is preferred to try/catch
a StepFailure than to specify this behavior as a variable.

BUG=chromium:727917
TEST=expectations

Change-Id: Ib3538a5a47d6a84258b7ce8c73062eaafdd70166