Fix toolchain packaging script for latest SDKs
On recent SDKs the size of the toolchain package grew. This change
filters out some unneeded directories in order to minimize this growth.

This change also fixes the bin paths for the 10.0.15063.0 SDK. Starting
with this SDK version the SDK version is part of the path. The script
does *not* handle packaging older SDKs anymore - older versions of the
script can be used for that purpose.

The 10.0.15063.0 SDK will be needed eventually in order to support
Windows 10 Creators Update.

Bug: 683729,682416
Change-Id: Ia89e3253869a45dd10c923a2edee53aaf086e12c