Update patch set 1
Patch Set 1:

(1 comment)

> I'd rather not add more confusion to this stuff, but if this is a path to getting rid of it, then I guess it's OK.

This isn't really getting rid of it, unfortunately. This is aimed at making it so that the toolchain can be installed / used in a directory other than "depot_toolS". On Chromium builds, we're going to create a special cache directory for the toolchain and load it there.

Patch-set: 1
Reviewer: Gerrit User 1109829 <1109829@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>