Update patch set 2
Patch Set 2:

I think this is my fault for not having considered how the GN exports would work across multiple deps files, so sorry about the confusion here.

The idea is that the exported args are then imported from an args file, which creates variables with global scope. As such, it doesn't really make sense to write values out to multiple files, and it's probably better to act more like we treat the default_args() scope in the GN dotfile, where only the outermost directive has any effect.

So, let's do this:

- Go back to using two separate args, but change the names to `gclient_gn_args` and `gclient_gn_args_file` for the list of args to export and the path, respectively.

- write the file out as just the list of args (not wrapped in a declare_args() function), so that it can be imported directly into an args.gn file or a default_args scope.

- Just put the GN arg-generating code directly into gclient, there's no real value in trying to use gn_helpers.py here and you're pulling in ~450 LOC to use < 30 of it. In particular, we might want to start by just supporting booleans. It's not clear if we need to support strings, and we probably don't want to support dicts or lists.

Patch-set: 2
CC: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>