Update patch set 4
Patch Set 4:

> > > Now that something has landed using this (I guess? My toolchain just redownloaded and is in c:\src\cr\src\third_party\depot_tools\win_toolchain\vs_files\3915730f76bd9c6155aed871b944b0a25c18f15e\... I realize we're going to orphan toolchains in the original location.
> > > 
> > > Before they were deleted if unused after 30 days, but that code isn't going to find the ones in the old root.
> > 
> > +dpranke for fyi.
> > 
> > I think once this settles we'll add a rule to cleanup_disk to nuke these on the bots. How much space are we talking about for devs? If necessary we could also add a hook to depot_tools to clean that up as well.
> 
> 1-3G.

Wait a minute, the description said that this changed *allowed* toolchains to be downloaded elsewhere, but it didn't sound like it would change the default location. Is the changelist description misleading?

Some developers may have a couple of copies of the toolchain (2015 and maybe one or more variants of VS 2017) so could have 2-6 GB (or much more in my case :-) but most will just have one.

Patch-set: 4