Update patch set 4
Patch Set 4:

> > Wait a minute, the description said that this changed *allowed* toolchains to be downloaded elsewhere, but it didn't sound like it would change the default location. Is the changelist description misleading?
> > 
> > Some developers may have a couple of copies of the toolchain (2015 and maybe one or more variants of VS 2017) so could have 2-6 GB (or much more in my case :-) but most will just have one.
> 
> I was wondering about that too. I saw this CL, but I couldn't find anything that actually started using it. Could someone confirm that we intentionally moved the toolchain location? (sometime on Tuesday MTV time I think, as it moved when I synced in the afternoon today.

I believe that this is correct - no change utilizing the new environment variable has been deployed, on developer systems or bots. I do have a CL staged that will utilize this, but it's not ready and certainly not landed.

I've been out for two days, so it's possible someone else took the initiative to implement this change without me being aware, but I doubt it, as the likely site of implementation seems to not include the requisite definition. You can confirm, I suppose, by pulling up a build and looking for the environment variable in the "runhooks" stage.

Patch-set: 4